requirements.txt
discord.py>=2.0.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
pytest>=6.0.0
pytest-asyncio>=0.15.0
psutil>=5.9.5
mplfinance>=0.12.7a0
pytz==2021.1
tensorflow>=2.15.0
keras==2.15.0
numpy==1.24.3
pandas==1.5.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.2.2
joblib==1.2.0
mysqlclient>=2.0.0
pymysql>=1.0.0
mysql-connector-python>=8.0.0
alembic>=1.7.0
prophet==1.1.6
xgboost==2.1.4
